@@ -3,23 +3,26 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:input_widgets_demo/inputs_demo/bloc/real_estate_search_form_cubit.dart';
 
 class CheckboxRow extends StatelessWidget {
-  const CheckboxRow({Key? key}) : super(key: key);
+  const CheckboxRow({super.key});
 
   @override
   Widget build(BuildContext context) {
     final isNewlyBuiltChecked =
         context.watch<RealEstateSearchFormCubit>().state.isNewlyBuiltChecked;
+
     return Row(
       children: [
         Checkbox(
           value: isNewlyBuiltChecked,
           onChanged: (checked) {
             if (checked != null) {
-              context.read<RealEstateSearchFormCubit>().newlyBuiltCheckedChanged(checked);
+              context
+                  .read<RealEstateSearchFormCubit>()
+                  .newlyBuiltCheckedChanged(checked);
             }
           },
         ),
-        const Text("Only newly built properties")
+        const Text("Only newly built properties"),
       ],
     );
   }
