@@ -5,7 +5,7 @@ import 'package:input_widgets_demo/inputs_demo/bloc/real_estate_search_form_cubi
 import '../../../util.dart';
 
 class SliderRow extends StatelessWidget {
-  const SliderRow({Key? key}) : super(key: key);
+  const SliderRow({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -28,9 +28,9 @@ class SliderRow extends StatelessWidget {
             ),
             onChanged: (values) {
               context.read<RealEstateSearchFormCubit>().onPriceValuesChanged(
-                    values.start,
-                    values.end,
-                  );
+                values.start,
+                values.end,
+              );
             },
           ),
         ),
