@@ -1,11 +1,10 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import '../main.dart';
 import 'second_page.dart';
 
 class FirstPage extends StatelessWidget {
-  const FirstPage({Key? key}) : super(key: key);
+  const FirstPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -16,23 +15,21 @@ class FirstPage extends StatelessWidget {
           padding: const EdgeInsets.all(16),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
+            spacing: 16,
             children: [
-              const Icon(
-                Icons.android,
-                color: Colors.greenAccent,
-                size: 120,
-              ),
-              const SizedBox(height: 16),
+              const Icon(Icons.android, color: Colors.greenAccent, size: 120),
               const Text("This is the first page"),
               ElevatedButton(
                 child: const Text("Navigate to the second page"),
                 onPressed: () {
-                  globalNavigatorHolderKey.currentState!.addNewPage(SecondPage(key: UniqueKey(),));
+                  globalNavigatorHolderKey.currentState!.addNewPage(
+                    SecondPage(key: UniqueKey()),
+                  );
                 },
               ),
               ElevatedButton(
                 child: const Text("Back"),
-                onPressed: (){
+                onPressed: () {
                   Navigator.of(context).pop();
                 },
               ),
