@@ -1,19 +1,15 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
 import '../data/photo_data_store.dart';
 
 class DetailedPage extends StatelessWidget {
-
-  const DetailedPage({Key? key}) : super(key: key);
+  const DetailedPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     final photo = ModalRoute.of(context)!.settings.arguments as StockPhoto;
     return Scaffold(
-      appBar: AppBar(
-        title: Text(photo.title),
-      ),
+      appBar: AppBar(title: Text(photo.title)),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         children: [
@@ -27,9 +23,7 @@ class DetailedPage extends StatelessWidget {
                   if (loading == null) {
                     return child;
                   } else {
-                    return const Center(
-                      child: CircularProgressIndicator(),
-                    );
+                    return const Center(child: CircularProgressIndicator());
                   }
                 },
               ),
@@ -39,21 +33,16 @@ class DetailedPage extends StatelessWidget {
             padding: const EdgeInsets.all(8.0),
             child: Text(
               photo.title,
-              style: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.w600,
-              ),
+              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.w600),
             ),
           ),
           Padding(
             padding: const EdgeInsets.symmetric(horizontal: 8.0),
             child: Text(
               photo.description,
-              style: const TextStyle(
-                fontSize: 14,
-              ),
+              style: const TextStyle(fontSize: 14),
             ),
-          )
+          ),
         ],
       ),
     );
