@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 
-import '../pages/detailed_page.dart';
 import '../data/photo_data_store.dart';
 
 class ListItemWidget extends StatelessWidget {
   final StockPhoto item;
 
-  const ListItemWidget({Key? key, required this.item}) : super(key: key);
+  const ListItemWidget({super.key, required this.item});
 
   @override
   Widget build(BuildContext context) {
@@ -18,23 +17,13 @@ class ListItemWidget extends StatelessWidget {
             aspectRatio: 1 / 1,
             child: Hero(
               tag: item.id,
-              child: Image.network(
-                item.url,
-                fit: BoxFit.cover,
-              ),
+              child: Image.network(item.url, fit: BoxFit.cover),
             ),
           ),
           title: Text(item.title),
-          subtitle: Text(
-            item.description,
-            maxLines: 1,
-          ),
+          subtitle: Text(item.description, maxLines: 1),
           onTap: () {
-            Navigator.pushNamed(
-              context,
-              "/details",
-              arguments: item,
-            );
+            Navigator.pushNamed(context, "/details", arguments: item);
           },
         ),
       ),
