@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'second_page.dart';
-
-
 class FirstPage extends StatelessWidget {
-  const FirstPage({Key? key}) : super(key: key);
+  const FirstPage({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -15,13 +12,9 @@ class FirstPage extends StatelessWidget {
           padding: const EdgeInsets.all(16),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
+            spacing: 16,
             children: [
-              const Icon(
-                Icons.android,
-                color: Colors.greenAccent,
-                size: 120,
-              ),
-              const SizedBox(height: 16),
+              const Icon(Icons.android, color: Colors.green, size: 120),
               const Text("This is the first page"),
               ElevatedButton(
                 child: const Text("Navigate to second page"),
