@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../main.dart';
 
 class SecondPage extends StatelessWidget {
-  const SecondPage({Key? key}): super(key: key);
+  const SecondPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -14,17 +14,16 @@ class SecondPage extends StatelessWidget {
           padding: const EdgeInsets.all(16),
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
+            spacing: 16,
             children: [
-              const FlutterLogo(
-                size: 120,
-              ),
-              const SizedBox(height: 16),
+              const FlutterLogo(size: 120),
               const Text("This is the second page"),
               ElevatedButton(
                 child: const Text("Navigate to the same page"),
                 onPressed: () {
-                  globalNavigatorHolderKey.currentState!
-                      .replaceTopPage(const SecondPage());
+                  globalNavigatorHolderKey.currentState!.replaceTopPage(
+                    const SecondPage(),
+                  );
                 },
               ),
               ElevatedButton(
